added clinvar coverage calculation for transcripts and moved some functions from views.py to clinvar_coverage.py file
